Change startup to use bootstrapApplication() with standalone component
