+++ src/app/app.component.ts
@@ -16,6 +16,7 @@ import { ThemeService } from './shared/services/theme/theme.service';
 import { Theme } from './shared/services/theme/theme';
 import { AppSettings } from './app-settings';
 import { LoadingSpinnerComponent } from './shared/loading-spinner/loading-spinner.component';
+import { TranslateModule } from '@ngx-translate/core';
 
 @Component({
   selector: 'bp-root',
